Update ShuntingYard.cpp
I used Simons rule to approximate an integral to get a numerical answer , initially I tried to iterate for small elements of dx and then sum each individually but this turns out to be computationally very heavy.
I came across various approximations but Simons rule seems to have the least error margin

The program works by using the shunting yard algorithm(as mentioned in the title) to parse the expression to postfix , after which we evaluate the postfix expression using a stack. The stack is evaluated in a nested for loop ,the outer loop being used to perform the sum required for each iteration of the integral approximation.